removed u_int32_t for windows

+++ src/kex.c
@@ -2123,8 +2123,8 @@ static int ecdh_sha2_nistp(LIBSSH2_SESSION *session, libssh2_curve_type type,
         /*host key K_S*/
         unsigned char *s = data + 1; //advance past packet type
         unsigned char *server_public_key;
-        u_int32_t server_public_key_len;
-        u_int32_t host_sig_len;
+        size_t server_public_key_len;
+        size_t host_sig_len;
 
         session->server_hostkey_len = _libssh2_ntohu32((const unsigned char*)s);
         s += 4;
